This is a group project and is a responsive web application using HTML, CSS, Bootstrap, JavaScript, and PHP. The Frontend part is completed. And also the Backend part is under construction now. It has a teacher and student login portal. The student can feedback to their teacher. The payment gateway is a working process.
This is a group project and is a responsive web application using HTML, CSS, Bootstrap, JavaScript, and PHP.
The Frontend part is completed. And also the Backend part is under construction now. It has a teacher and
student login portal. The student can feedback to their teacher. The payment gateway is a working process.